pyproject.toml
[tool.poetry]
name = "django-practice"
version = "0.1.0"
description = ""
authors = ["Jesse He <hjesse92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.1"
djangorestframework = "^3.14.0"
celery = "^5.3.0"
redis = "^4.5.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
